requirements.txt
requests
ipykernel
python-dotenv
openai
pydantic
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
